requirements.txt
Django==1.8.4
django-wordpress-auth-lite==0.1.0
MySQL-python==1.2.5
phpserialize==1.3
wheel==0.24.0
